[WIP] move BaseTest funtionality to .cpp file
