+++ src/test/lib/import_export/binary/binary_writer_test.cpp
@@ -23,30 +23,6 @@ class BinaryWriterTest : public BaseTest {
 
   void TearDown() override { std::remove(filename.c_str()); }
 
-  bool file_exists(const std::string& name) {
-    std::ifstream file{name};
-    return file.good();
-  }
-
-  bool compare_files(const std::string& original_file, const std::string& created_file) {
-    std::ifstream original(original_file);
-    Assert(original.is_open(), "compare_file: Could not find file " + original_file);
-
-    std::ifstream created(created_file);
-    Assert(created.is_open(), "compare_file: Could not find file " + created_file);
-
-    std::istreambuf_iterator<char> iterator_original(original);
-    std::istreambuf_iterator<char> iterator_created(created);
-    std::istreambuf_iterator<char> end;
-
-    while (iterator_original != end && iterator_created != end) {
-      if (*iterator_original != *iterator_created) return false;
-      ++iterator_original;
-      ++iterator_created;
-    }
-    return ((iterator_original == end) && (iterator_created == end));
-  }
-
   std::shared_ptr<Table> table;
   const std::string filename = test_data_path + "export_test.bin";
   const std::string reference_filepath = "resources/test_data/bin/";
