+++ src/test/operators/export_test.cpp
@@ -1,5 +1,4 @@
 #include <cstdio>
-#include <fstream>
 #include <memory>
 #include <string>
 #include <utility>
@@ -35,30 +34,6 @@ class OperatorsExportTest : public BaseTest {
   void TearDown() override {
     std::remove(test_filename.c_str());
     std::remove(test_meta_filename.c_str());
-  }
-
-  bool file_exists(const std::string& name) {
-    std::ifstream file{name};
-    return file.good();
-  }
-
-  bool compare_files(const std::string& original_file, const std::string& created_file) {
-    std::ifstream original(original_file);
-    Assert(original.is_open(), "compare_file: Could not find file " + original_file);
-
-    std::ifstream created(created_file);
-    Assert(created.is_open(), "compare_file: Could not find file " + created_file);
-
-    std::istreambuf_iterator<char> iterator_original(original);
-    std::istreambuf_iterator<char> iterator_created(created);
-    std::istreambuf_iterator<char> end;
-
-    while (iterator_original != end && iterator_created != end) {
-      if (*iterator_original != *iterator_created) return false;
-      ++iterator_original;
-      ++iterator_created;
-    }
-    return ((iterator_original == end) && (iterator_created == end));
   }
 
   std::shared_ptr<Table> table;
